Add support for additional plugin directories in qsgen2
This commit modifies the qsgen2 script to add support for additional plugin directories. The fpath variable is updated to
include /home/stig/.config/qsgen2/plugins in addition to /home/stig/bin/include/common. This allows users to place their custom
plugins in the ~/.config/qsgen2/plugins directory and have them loaded by the script.

Additionally, a debug message is added to notify the user if no valid generator is found in the 'config' file, and the
script exits in that case.

The commit also includes some minor code formatting and comment updates for clarity.