Refactor _blogs and _blog_index functions
This commit refactors the _blogs and _blog_index functions in the qsgen2 script.

In the _blogs function, the directory creation message is updated to use the correct variable  instead of . Additionally, a debug message is added to indicate when  is set to true.

In the _blog_index function, debug messages are added to display the values of  and . The debug message for generating the page is updated to include the correct file path .

These changes improve the clarity and accuracy of the script's functionality.