Fix debug flag in _blogs function and add debug messages in _blog_index function
In the _blogs function, the debug flag was set to true, which caused unnecessary debug output. Changed it to false.

In the _blog_index function, added debug messages to indicate the replacement of sitename and tagline, as well as the replacement of BODY with the content of the blog index file.

This commit fixes the debug flag and adds debug messages for better troubleshooting and understanding of the code.