Add blog index content to /blog/index.html
This commit adds the blog index content to the file /blog/index.html. The function  replaces the  placeholder in the content with the actual blog index list. After the replacement, the updated content is written to the file /blog/index.html. The commit also includes debug messages to display the content of  before writing it to the file.