Refactor code in qsgen2 script
This commit refactors the code in the qsgen2 script by removing unnecessary commented-out lines and cleaning up variable assignments. It also improves the readability of the code by using conditional assignment for the  variable. Instead of assigning the value of  directly, it checks if  is set to true and assigns the corresponding boolean value to . This simplifies the code and makes it easier to understand.